V.1.3.1 Test Git hub
This is a test for my main code. This test Included Changed API with changed stat card, code does not include my filter , card is not currently displaying info on card, I need to use v.1.2.3 for stat card to display movie information. So var V.1.2.3 might be better for the group work, i will not continue with this version if the team does not liked this mixed code. 

compared to old code:
filter is on V.1.2.3 
Working card on V.1.2.3 
New API V.1.3.1